services/lpr_mercosul/requirements.txt
opencv-python-headless==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.0
imutils==0.5.4
scikit-image==0.21.0
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu
ultralytics==8.0.196
